send On branch circle_2 Your branch is up to date with origin/circle_2. 	modified:   circle_2/_push_swap/Makefile 	modified:   circle_2/_push_swap/libstack/stack.h 	modified:   circle_2/_push_swap/libstack/stack_utils.c 	modified:   circle_2/_push_swap/push_swap.c 	modified:   circle_2/_push_swap/sorting_algo/algo.h 	modified:   circle_2/_push_swap/utils/error.c 	modified:   circle_2/_push_swap/utils/start.c 	modified:   circle_2/_push_swap/utils/utils.h 	circle_2/_push_swap/sorting_algo/sort.c to github

+++ circle_2/_push_swap/libstack/stack.h
@@ -6,7 +6,7 @@
 /*   By:  mchenava < mchenava@student.42lyon.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/08 09:37:23 by  mchenava         #+#    #+#             */
-/*   Updated: 2023/01/04 10:27:44 by  mchenava        ###   ########.fr       */
+/*   Updated: 2023/01/09 10:08:17 by  mchenava        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_node
 {
-	int				value;
+	long int		value;
 	int				position;
 	struct s_node	*next;
 	struct s_node	*prev;
